Fixed errors in tests
